+++ dev/reports/experiments_2023-02-xx.sh
@@ -2449,9 +2449,9 @@ python -m watch.mlops.aggregate \
         stats_ranking: 0
         min_variations: 1
         params_of_interest:
-            - sv_depth_filter.threshold
-            - sv_depth_score.model_fpath
-            - bas_poly.thresh
+            - params.sv_depth_filter.threshold
+            - params.sv_depth_score.model_fpath
+            - params.bas_poly.thresh
     " \
     --stdout_report="
         top_k: 10
@@ -2461,7 +2461,8 @@ python -m watch.mlops.aggregate \
         print_models: True
         reference_region: final
     " \
-    --rois="KR_R002,PE_R001,NZ_R001,CH_R001,KR_R001,AE_R001,BR_R002,BR_R004"
+    --rois="PE_R001"
+    #--rois="KR_R002,PE_R001,NZ_R001,CH_R001,KR_R001,AE_R001,BR_R002,BR_R004"
 
     #--rois="KR_R002"